+++ TelegramBotAPI/src/jvmMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/utils/extensions/Webhooks.kt
@@ -29,7 +29,7 @@ import java.util.concurrent.TimeUnit
  * which will be used by telegram to send encrypted messages
  * @param scope Scope which will be used for
  */
-@Deprecated("Replaced into project TelegramBotAPI-extensions-api")
+@Deprecated("Replaced into project TelegramBotAPI-extensions-utils")
 suspend fun RequestsExecutor.setWebhook(
     url: String,
     port: Int,
@@ -137,7 +137,7 @@ suspend fun RequestsExecutor.setWebhook(
     }
 }
 
-@Deprecated("Replaced into project TelegramBotAPI-extensions-api")
+@Deprecated("Replaced into project TelegramBotAPI-extensions-utils")
 suspend fun RequestsExecutor.setWebhook(
     url: String,
     port: Int,
@@ -165,7 +165,7 @@ suspend fun RequestsExecutor.setWebhook(
     block
 )
 
-@Deprecated("Replaced into project TelegramBotAPI-extensions-api")
+@Deprecated("Replaced into project TelegramBotAPI-extensions-utils")
 suspend fun RequestsExecutor.setWebhook(
     url: String,
     port: Int,
@@ -190,7 +190,7 @@ suspend fun RequestsExecutor.setWebhook(
     block
 )
 
-@Deprecated("Replaced into project TelegramBotAPI-extensions-api")
+@Deprecated("Replaced into project TelegramBotAPI-extensions-utils")
 suspend fun RequestsExecutor.setWebhook(
     url: String,
     port: Int,